+++ R/runcode.R
@@ -7,13 +7,11 @@
 #' and while developing an app locally, but it \strong{should not be included in
 #' an app that is accessible to other people}, as letting others run arbitrary R
 #' code can open you up to security attacks.\cr\cr
-#' To use this construct, you must add a call to \code{runcodeUI()} in the UI
-#' of your app, and a call to \code{runcodeServer()} in the server function. You
-#' also need to initialize shinyjs with a call to \code{useShinyjs()} in the UI.
+#' To use this construct, you must add a call to \code{runcode()} in the UI
+#' of your app, and initialize shinyjs with a call to \code{useShinyjs()}.
 #'
-#' @note You can only have one \code{runcode} construct in your shiny app.
-#' Calling this function multiple times within the same app will result in
-#' unpredictable behaviour.
+#' @note You can only have one `runcode` construct in your shiny app, do not
+#' call it more than once in the UI.
 #'
 #' @param code The initial R code to show in the text input when the app loads
 #' @param type One of \code{"text"} (default), \code{"textarea"}, or \code{"ace"}.
@@ -26,8 +24,7 @@
 #' \code{type="ace"})
 #' @param height The height of the editable code input (ignored when
 #' \code{type="text"})
-#' @param ns When used inside a shiny module, the module's \code{\link[shiny]{NS}}
-#' needs to be provided. This argument should remain \code{NULL} when not used inside a module.
+#' @param ns The [`namespace`][shiny::NS()] object of the current module if inside a Shiny module.
 #' @seealso \code{\link[shinyjs]{useShinyjs}}
 #' @examples
 #' if (interactive()) {
@@ -36,22 +33,17 @@
 #'   shinyApp(
 #'     ui = fluidPage(
 #'       useShinyjs(),  # Set up shinyjs
-#'       runcodeUI(code = "shinyjs::alert('Hello!')")
+#'       runcode(code = "shinyjs::alert('Hello!')")
 #'     ),
-#'     server = function(input, output) {
-#'       runcodeServer()
-#'     }
+#'     server = function(input, output) {}
 #'   )
 #' }
-#' @name runcode
-
-#' @rdname runcode
 #' @export
-runcodeUI <- function(code = "",
-                      type = c("text", "textarea", "ace"),
-                      width = NULL,
-                      height = NULL,
-                      ns = shiny::NS(NULL)) {
+runcode <- function(code = "",
+                    type = c("text", "textarea", "ace"),
+                    width = NULL,
+                    height = NULL,
+                    ns = shiny::NS(NULL)) {
   type <- match.arg(type)
 
 	if (type == "ace") {
@@ -61,57 +53,57 @@ runcodeUI <- function(code = "",
 	}
 
   placeholder <- "Enter R code"
-  shiny::singleton(shiny::tagList(
-    if (type == "text")
+  id <- ns(paste0("runcode_expr"))
+  codeTag <-
+    if (type == "text") {
       shiny::textInput(
-        ns("runcode_expr"), label = NULL, value = code,
+        id, label = NULL, value = code,
         width = width, placeholder = placeholder
-      ),
-    if (type == "textarea")
+      )
+    } else if (type == "textarea") {
       shiny::textAreaInput(
-        ns("runcode_expr"), label = NULL, value = code,
+        id, label = NULL, value = code,
         width = width, height = height, placeholder = placeholder
-      ),
-    if (type == "ace")
-			shinyAce::aceEditor(ns("runcode_expr"), mode = 'r', value = code,
-				height = height, theme = "github", fontSize = 16),
-    shiny::actionButton(ns("runcode_run"), "Run", class = "btn-success"),
+      )
+    } else if (type == "ace") {
+      shinyAce::aceEditor(id, mode = 'r', value = code,
+                          height = height, theme = "github", fontSize = 16)
+    }
+
+  callback <- paste0(
+    "Shiny.setInputValue(",
+    "'runcode:shinyjspkg',",
+    "{ ns : '", ns(""), "', code : document.getElementById('", id, "').value },",
+    "{ priority : 'event' }",
+    ");")
+
+  shiny::tagList(
+    codeTag,
+    shiny::tags$button("Run", onclick = HTML(callback), class = "btn btn-success action-button"),
     shinyjs::hidden(
       shiny::div(
-        id = ns("runcode_error"),
+        id = ns(paste0("runcode_error")),
         style = "color: red; font-weight: bold;",
         shiny::div("Oops, that resulted in an error! Try again."),
         shiny::div("Error: ", shiny::br(),
                    shiny::tags$i(shiny::span(
-                     id = ns("runcode_errorMsg"),
+                     id = ns(paste0("runcode_errorMsg")),
                      style = "margin-left: 10px;")))
       )
     )
-  ))
+  )
 }
 
-#' @rdname runcode
+#' Deprecated - use `runcode()` instead
+#' @inheritParams runcode
+#' @export
+runcodeUI <- function(...) {
+  warning("shinyjs: `runcodeUI()` is now deprecated in favour of `runcode()`")
+  runcode(...)
+}
+
+#' Deprecated - use `runcode()` instead
 #' @export
 runcodeServer <- function() {
-  # evaluate expressions in the caller's environment
-  parentFrame <- parent.frame(1)
-
-  # get the Shiny session
-  session <- getSession()
-
-  shiny::observeEvent(session$input[['runcode_run']], {
-    shinyjs::hide("runcode_error")
-
-    tryCatch(
-      shiny::isolate(
-        eval(parse(text = session$input[['runcode_expr']]), envir = parentFrame)
-      ),
-      error = function(err) {
-        shinyjs::html("runcode_errorMsg", as.character(err$message))
-        shinyjs::show(id = "runcode_error", anim = TRUE, animType = "fade")
-      }
-    )
-  })
-
-  invisible(NULL)
+  warning("shinyjs: `runcodeServer()` is now deprecated. You no longer need to call any function in the server, simply call `runcode()` in the UI.")
 }