+++ inst/examples-shiny/sandbox/server.R
@@ -10,6 +10,4 @@ shinyServer(function(input, output, session) {
     onclick(sprintf("example-%s", names(ex)),
             updateTextInput(session, "runcode_expr", value = as.character(ex)))
   })
-
-  runcodeServer()
 })